org.Osativa.eg.db {org.Osativa.eg.db}	R Documentation

Bioconductor annotation data package

Description

Welcome to the org.Osativa.eg.db annotation Package. This is an organism specific package. The purpose is to provide detailed information about the species abbreviated in the second part of the package name org.Osativa.eg.db. This package is updated biannually.

To learn more about this package, users are encouraged to learn about the select, columns, keys and keytypes methods. These are described in a walkthrough on the bioconductor website as well as in the manual pages and vignettes in the AnnotationDbi package.

[Package org.Osativa.eg.db version 1.0 Index]